BrdCfg_1RENA_50MHzSMA: Termination Resistor...
Many changes needed to add a termination resistor to the LVDS
clock input. Had to manually place a buffer between the top level
port and the PLL, because the PLL's automatically-generated
differential input does not generate code with a terminating
resistor. So I seperated the differential buffer into a seperate
part.