RENA_FPGA: Add crc8 to tx, and rearrange tx.
Add a crc8 to the end of all outgoing packets, before the final x"FF".
It is split into two 4-bit parts, due to 6 bit limitations of packet
data.

Changed the state machine around in the RS232_tx_buffered component,
to make the code more understandable and compatible with the
crc8. Previously there wasn't a good way to insert bytes into
the packet randomly.