Add FPGA_Shared/sapet_packets.vhd. Use all around.
Adds various constants and functions that are used to identify
properties of packets such as start/end tokens and maximum
packet length. These constants are replaced throughout, with
the exception of certain bug-finding code and the
end_of_packet_token, which was not replaced everwhere due
to the massive number of them.