Backplane_FPGA: Overhaul Acquisition_module.
Re-did lots of stuff.

Took the FIFO related portions out and replaced
them with a stripped down version of the smart packets fifo, called
packets_fifo_1024_16. The difference with the smart packet fifo
is that this does not include any logic to expose the source or
destination node addresses, which simplifies the component considerably.

The processes which decode the serial data into parallel format were
placed into a component called the deserializer. They were mostly
unchanged, except for untangling with other code.

The code which switched between the 2 input ports was replaced
with code that is essentially a stripped down version of the new
Daisychain_module routing code.

Should now function almost identically from a blackbox perspective.