Backplane_FPGA/05_acquisition: Add input chooser.
In the aquisition module, which currently inputs from two sources,
remove the previously added logic for choosing between which
input source to use, and create a new component for the task, the
input_chooser_2_sources. This component should be closer to what
is needed for a recursive tree to connect 48 frontend boards.