Backplane_FPGA/05_acquisition: Add synchronizer.
Added the packet_bus_output_synchronizer to the
input_shooer_2_sources. This makes the output synchronous,
without loss of any performance or usability, just one cycle
of propogation delay from input interface to output interface.
There is no delay between input/outputs of the output interface,
so usage is unchanged.

This is intended so that we can cascade input switches without
fear of breaking timings.