Backplane_FPGA/BrdCfg_48RENA...: 50MHz Clock tweak
The system now runs off of the 50MHz clock output from the PLL,
instead of from the external source 50MHz clock directly. This
seems to fix an inability to make correct timing. I hypothesize that
this signal is buffered better, or placed better, or has less
skew relative to other clocks? Not really sure why but it seemed
like a good idea and fixed the problem.

If I recall correctly, the timing error actually came from something
in the UDP?