Backplane_FPGA/.../input_fifo_switch: Fix loop issue.
Add a port, dout_empty_notready_source_last_tick, which is the same
as dout_empty_notready except it gives the empty_notready signal for
the input selected in the previous cycle (the empty_notready signal
is current, the selected input is the previous). This is useful
to break combinational loops, etc, that can feed in through the
input selection bits. In the middle of a packet, we don't need
to change sources, so can use this new signal.