Backplane_FPGA: Now does 48 sources (4 connected).
input_choosers_N_sources now works with an arbitrary number of sources.
Connected the 48 boards project up for all the pins, but only frontend
board receivers 0-3 are connected in the top level module.

Also had to hack the constraints file for the 48 board project to
think the pins are 2.5 volts. See comment at top of constraint file.