Backplane_FPGA: Fifo full corruption/handler LEDs
Add detection of a condition where the FIFO overflows because
a packet came in that was larger than the specified max packet
size. This would be a bug/malfunction condition, but previously
it would permanently lock up the FIFO.  Now it is handled in a more
friendly manner, and LEDs are hooked up to blink when this happens.