Backplane_FPGA: Heed the GTP out busy.
The main daisychain module will halt when trying to write to the
GTP but the GTP is flagging that it is busy. In practice I don't
think this ever happens, but, I will sleep a bit better at night
knowing this is here. Also has provisions for a busy signal on
any output channel (ethernet, frontend configuration serializer,
etc).